make healpix diagram pretty
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@8713 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/healpix2.py
@@ -27,7 +27,9 @@ _hp = ctypes.CDLL(os.path.join(os.path.dirname(__file__), '_healpix.so'))
 def healpix_to_radecdeg(hp, nside=1., dx=0.5, dy=0.5):
     ra = ctypes.c_double(0)
     dec = ctypes.c_double(0)
-    _hp.healpix_to_radecdeg(hp, nside, dx, dy,
+    _hp.healpix_to_radecdeg(hp, nside,
+							c_double(dx),
+							c_double(dy),
                             ctypes.pointer(ra),
                             ctypes.pointer(dec))
     return (ra.value, dec.value)
